Refactor: Loading.tsx를 app 디렉토리로 이동
